[Flow] Support dispatch tracing when encodings are present (#20447)
Enables dispatch tracing when the dispatch op has encoded operands or
results. New `flow.tensor.encode` ops are inserted to bring the tensor
back to the row-major layout before adding the tracing ops because the
`flow.tensor.trace` op cannot handle encoded tensors. The tensor is not
printed in the encoded layout with this revision, but it still adds
helpful debug information, and prevents compilation failures when adding
dispatch tracing.

---------

Signed-off-by: Max Dawkins <max.dawkins@gmail.com>