[DT] Split out iree_encoding.layout from EncodingAttr. (#20507)
The revision implements `IREE::Encoding::LayoutAttr` and retires the
`layouts` field from EncodingAttr, as it encodes much more information
than needed.

It is not easy to do incremental because many layers are involved,
including
- Encoding specialization.
- Encoding materialization.
- Lower `iree_codegen.query_tile_size` op for VMVX ukernel path.

To make review easier, there are five commits in the PR:
- Introduce iree_encoding.layout encoding attribute:
https://github.com/iree-org/iree/commit/85d919941ebc9ea1fa17de77e2716d3f332236f4
- Implement SerializableEncodingInterface for Encoding::LayoutAttr:
https://github.com/iree-org/iree/commit/a49a7263b471f8f8112e083b53e97540a6cf70b3
- Use iree_encoding.layout when the layouts are resolved:
https://github.com/iree-org/iree/commit/1dc9007118a27ec32508277944acad373760a96c
- Retire the layouts field from EncodingAttr:
https://github.com/iree-org/iree/commit/b2f0329047c4c819dd9c0fa85581a933c99969b3
- Fix the ukernel ops lowering for VMVX:
https://github.com/iree-org/iree/commit/0a51344b21512329c3b82ba74ff696241ce0cf6f

In the third commit, we need to move `registerAttributes` to
`EncodingAttr.cpp` because we want to skip default builders.
It is very tricky when the verifier is implemented, because an assertion
is triggered. However, it should just return nullptr in this context. To
bypass the crash in get method, we need to implement our own get
methods.

Note: the root cause of `reigsterAtrributes` is that the full definition
of the storage class needs to be visible whenever we register the
attribute.

Fixes https://github.com/iree-org/iree/issues/20489

---------

Signed-off-by: hanhanW <hanhan0912@gmail.com>