[Codegen] Add pass for interpreting transform specs on lowering configs (#20408)
The existing mechanism for externally authored codegen strategies via
transform dialect overrides almost the entire codegen pipeline with the
transform dialect strategy. This does not scale as it typically either
a) Is no better than a standard pass pipeline, in which case there is no
      reason to use transform dialect.
or b) Is specialized for a single case which railroads future graph
      level/fusion efforts.

A more scalable middle ground is to thread through transform strategies
that apply specifically to certain operations and let the surrounding
pipeline handle the rest. This patch adds a pass that walks a module and
looks up + applies a transform strategy on any annotated op.

One concern with transform dialect passes has been dialect registration.
The only dialect this pass *requires* is iree_codegen to load the
strategy library. Otherwise the current assumption is that this will be
used within a TargetBackend's pass pipeline, meaning each target backend
adds the dialect that needs to be registered as a part of the
surrounding pipeline. I'm not 100% how this will pan out with
reproducers but there aren't any good alternatives (registering
everything is a no-go).