[Im2col] Add input-filter permutation info to im2col metadata (#20531)
This PR extends the `im2col` operation to support a new attribute
`input_filter_perm`, which encodes the permutation required to align the
layout of input and filter reduction dimensions. This is useful when the
logical layout of the filter (e.g., `CHW`) differs from that of the
input (e.g., `HWC`). Then in im2col decomposition, using permutation
metadata to handle layout mismatches and reorder index computations.

This fixes https://github.com/iree-org/iree/issues/20473.

---------

Signed-off-by: yzhang93 <zhyuhang88@gmail.com>