[TensorExt][NFC] Remove dependency on Encoding through ExternalModel (#20612)
There is no strong dependency of `DispatchTensorType` on the Encoding
dialect as the encoding can be just any attribute and is optional, so I
don't think there's a need to have the TensorExt dialect explicitly
depend on the Encoding just for the interface. Decoupling them would
avoid potential circular dependency issues (I ran into that while
working on
https://github.com/iree-org/iree/issues/20160#issuecomment-2756081050),
remove the need for other (potential) users of the TensorExt dialect to
use the encoding and allow users to bring their own encoding interfaces
while avoiding having to bring in the Encoding dialect.

Signed-off-by: Jorn Tuyls <jorn.tuyls@gmail.com>