e2e matmul tests: log early the presence of a numerical error (#20638)
On numerical error, e2e matmul tests rerun the reference matmul on ALL
elements (instead of computing only 1 every N elements) to print the
detailed numerical diagnostic. That can take a minute on huge matrices,
as the reference code is so slow. Logging early what is going on allows
the user to know it's a numerical error and not a hang.

Signed-off-by: Benoit Jacob <jacob.benoit.1@gmail.com>