[LinalgExt] Add map_scatter op and verifier (#20640)
Adds a new op called `iree_linalg_ext.map_scatter`, which represents a
scattering of an input tensor to an output tensor by a mapping from
input indices to unique output indices. The index mapping is represented
by a region containing index computation operations, yielding a set of
output indices and a mask.

The new op is useful for representing long chains of index
transformation in a single operation, and can greatly simplify the
handling of complex sequences of index transformation ops, like we often
see in data tiling layout transformations.

This PR adds the op definition and verifier for the new op.

---------

Signed-off-by: Max Dawkins <max.dawkins@gmail.com>