[DispatchCreation] Set padding encodings on intermediate tensors. (#20634)
This PR adds a new way to set encodings to suggest that certain
intermediate tensors need to have a wider allocation than the logical
shape of the tensor.

In this implementation, the lhs/rhs operands of contraction-like
operations that are produced by other dispatches are targets for adding
such encoding. This PR also accounts for having operations like
`tensor.collapse_shape`/`tensor.expand_shape` in between the producing
dispatch and its use. It doesnt do this through propagation, instead,
the intervening operations are cloned into the producer dispatch, so
that the result of the producer dispatch is directly usable in the
operation. The result of the dispatch is then "encoded" to specify that
padding is needed.

Depends on #20662 

Please review only last commit.

Signed-off-by: MaheshRavishankar <mahesh.ravishankar@gmail.com>