Runtime float type conversion helpers: Fix handling of denormals. (#20676)
As preparation for introducing support for FP6/FP4 types where denormals
become essential, I wanted to make sure that we had robust support for
denormals already for larger types. It was known to be incomplete in the
sense that we were still flushing to zero in many cases, but it turned
out to also be buggy in some cases. This PR should nail this for good.

---------

Signed-off-by: Benoit Jacob <jacob.benoit.1@gmail.com>