[Codegen][DerivedConfig] Add support to set outermost tile size as vector size (#20692)
In backward convolutions (e.g., `conv2d_chwn_chwf`), the im2col input
tensor has the batch dim as innermost, whereas the output tensor
implicitly transposes the batch dimension to the outermost. To enable
efficient `vector.load` operations on the im2col input, we need to tile
along the contiguous batch dimension and assign the vector size to the
outermost tile dimension.

This PR together with https://github.com/iree-org/iree/pull/20633 will
improve the performance of backward convolutions. For example, for the
following conv
```
%7 = linalg.generic {indexing_maps = [affine_map<(d0, d1, d2, d3, d4, d5, d6) -> (d4, d1 + d5, d2 + d6, d3)>, affine_map<(d0, d1, d2, d3, d4, d5, d6) -> (d4, d5, d6, d0)>, affine_map<(d0, d1, d2, d3, d4, d5, d6) -> (d0, d1, d2, d3)>], iterator_types = ["parallel", "parallel", "parallel", "parallel", "reduction", "reduction", "reduction"]} ins(%3, %4 : tensor<16x26x18x288xbf16>, tensor<16x24x16x288xbf16>) outs(%6 : tensor<288x3x3x288xf32>) {
    ^bb0(%in: bf16, %in_0: bf16, %out: f32):
      %10 = arith.extf %in : bf16 to f32
      %11 = arith.extf %in_0 : bf16 to f32
      %12 = arith.mulf %10, %11 : f32
      %13 = arith.addf %out, %12 : f32
      linalg.yield %13 : f32
    } -> tensor<288x3x3x288xf32>
```
The runtime is reduced from 235us to 140us.

---------

Signed-off-by: yzhang93 <zhyuhang88@gmail.com>