Cleaning up iree_hal_module_debug_sink_t destroy. (#20841)
It was not properly added to the inline HAL, did not handle the status,
and was misnamed for what it was doing (destroy = delete). It's still
not great that this grew this method but it was confusing as it was. If
debug sinks need lifetime management they should be promoted to a full
ref counted type instead of an on-stack by-value struct as was the
original intent.