[Integrate] Make IREE compatible with the new memref.assume_alignment semantic change. (#20913)
This drops a revert:
https://github.com/iree-org/llvm-project/commit/435bb5041f83d11996fe9c47c38288cb177d3dda

To make IREE compatible with the new `memref.assuem_assignment`
semantics, the revision moves the op creation from OneShotBufferization
to post bufferization passes. It is mainly for reducing the burdens of
bufferization analysis. Otherwise, it'd easily generate redundant copies
between the original memref and the result of AssumeAlignment op.

Furthermore, there are several missing support when AssumeAlignment op
is present in SSA chain.
- The `FlattenMemRefSubspanPass` needs to be improved to take the op
into account.
- We are missing a pattern that propagates the `memref.dim` ops for
`memref.dim(memref.assume_alignment)` cases. It reduces the deps from
AssumeAlignment op, so we don't need to add more patterns for it.
- The chained memref.assume_alignment ops are not folded away. This can
be done as a canonicalization pattern. The revision adds a workaround
for early-bufferization dispatch, and it will be removed once we
implement the pattern.
- Need to teach SPIRVVectorizeLoadStore pass to take AssumeAlignment ops
into account.
- Need to teach VMVX ResolveBufferDescriptors pass to take
AssumeAlignment ops into account. The new pattern is copied from the
other one, and a refactoring can be done in the future. For now, we
mainly want to drop the revert, and we can easily refactor the code out.
So it is a TODO for now.
- It triggers failures in hoisting vector transfer ops when
AssumeAlignment ops are present. We don't know the root cause yet, and
it is too complicated and hard to fix it within the revision. Because
the implementation is in upstream. See
https://github.com/iree-org/iree/issues/20912 for reproducer.

Closes https://github.com/iree-org/iree/issues/20908

---------

Signed-off-by: hanhanW <hanhan0912@gmail.com>