Refresh the uses of memref.assume_alignment in lit tests. (#20925)
Few tests are updated to use the result of `memref.assume_alignment`
because they care about the ops.

Many `memref.assume_alignment` ops are removed from many lit tests
because they do not need them. They don't use them at their pass scope.

There are two tests not changed because of recent regression. I'll
revisit it if I come back to the issue. For now, leave them as what they
are:
- SPIRV/test/tile_and_promote_cooperative_matrix.mlir
- SPIRV/test/tile_and_vectorize_to_cooperative_ops.mlir

This is a follow-up to
https://github.com/iree-org/iree/commit/a8983ee52407c60b55857ddaf274e2a3350c2319

---------

Signed-off-by: hanhanW <hanhan0912@gmail.com>