Integrate llvm-project@d45031ce5281 (#20924)
Dropped reverts, which are fixed by the head:
-
https://github.com/iree-org/llvm-project/commit/0357fd9d0583393135ecb651b038acd39d7c9208
-
https://github.com/iree-org/llvm-project/commit/20a9a98f1430384fcd602ce560a36f93c6cd1b2d

Carried revert:
- (old)
https://github.com/iree-org/llvm-project/commit/13631cf47007968f6790c4cdf39b8c3784d819e9
- (new)
https://github.com/iree-org/llvm-project/commit/5f144b5d8134c23f4028d5e7a334d40b6f260252
- (new)
https://github.com/iree-org/llvm-project/commit/cddcc131aace5714f819b2189dd22bc0f10427d4
- (new)
https://github.com/llvm/llvm-project/commit/989aadf4b112ef55648d801c4003b63f8aad930e

Additional cherry-picks, which can be dropped after we bump ahead of it:
-
https://github.com/iree-org/llvm-project/commit/92b96897b0edd124934daddf96f3468d488924d6

The fixes in IREE are for
-
https://github.com/llvm/llvm-project/commit/05494f3bad0322f2f0d2128857ba30479d97deb2
-
https://github.com/llvm/llvm-project/commit/61d5fdf50c78810972f5984473600fd917ccaff5

The second and the third reverts because they break torch-mlir. I have a
fix for IREE side and can fix torch-mlir together in our repo. I prefer
fixing them later because we are already ~1 week behind the LLVM head.

The four revert is because of an assertion in convert-to-llvm lowering
in ARM SME pipeline tests.

Fixes https://github.com/iree-org/iree/issues/20737

---------

Signed-off-by: hanhanW <hanhan0912@gmail.com>